Add config to keep initial noise during validation (#101)
* Allow train.py to use the same seed during validation

Clones a copy of the initial random noise to CPU when deterministic_validation is True. This noise is then used during training validation and is equivalent to supplying the same RNG seed for each cycle.

* Add deterministic_validation to base config PixArt_xl2_internal.py

Disabled by default to retain expected behavior.

* Add missing else None to deterministic_validation

* Remove extraneous argument and use proper None check

The namespace available included the value being passed in the parameters. In addition, Python prefers explicit checks for None.

* Fix old variable reference